chore(e2e): move query rules
