+++ e2e/setup/scenarios.ts
@@ -2,5 +2,6 @@ import './scenarious.world';
 
 import './scenarios.env';
 import './scenarios.wait';
-import './scenarios.query';
 import './scenarios.screenshot';
+
+import './element/scenarios.query';
